requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
attrs==24.3.0
Automat==24.8.1
beautifulsoup4==4.12.3
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.2
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.0
cssselect==1.2.0
cssselect2==0.8.0
defusedxml==0.7.1
filelock==3.16.1
Flask==3.1.1
fonttools==4.59.0
frozenlist==1.5.0
geopandas==1.1.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.25.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.10.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
lxml==5.3.0
Markdown==3.8.2
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
parsel==1.9.1
pefile==2023.2.7
pillow==11.3.0
propcache==0.2.1
Protego==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyDispatcher==2.0.7
pydyf==0.11.0
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.8
pyogrio==0.11.0
pyOpenSSL==24.3.0
pyparsing==3.2.3
pyphen==0.17.2
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
pywin32-ctypes==0.2.3
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
rsa==4.9.1
Scrapy==2.12.0
service-identity==24.2.0
setuptools==75.6.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
tenacity==8.5.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tldextract==5.1.3
tqdm==4.67.1
Twisted==24.11.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.2.0
urllib3==2.3.0
w3lib==2.2.1
weasyprint==66.0
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.18.3
zope.interface==7.2
zopfli==0.2.3.post1
